// app/src/generated/models/Ben_ltkpriorityassignmentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Ben_ltkpriorityassignmentsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Ben_ltkpriorityassignmentsstatecode = keyof typeof Ben_ltkpriorityassignmentsstatecode;
export const Ben_ltkpriorityassignmentsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Ben_ltkpriorityassignmentsstatuscode = keyof typeof Ben_ltkpriorityassignmentsstatuscode;

export interface Ben_ltkpriorityassignmentsBase {
  ben_area?: string;
  ben_childpriorityid?: string;
  ben_company?: string;
  ben_decidedat?: string;
  ben_decidedbyid?: string;
  ben_decidedbyname?: string;
  ben_department?: string;
  ben_ltkpriorityassignmentid: string;
  ben_name?: string;
  "ben_Priority@odata.bind"?: string;
  ben_reason?: string;
  ben_site?: string;
  ben_status?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Ben_ltkpriorityassignmentsstatecode;
  statuscode?: Ben_ltkpriorityassignmentsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Ben_ltkpriorityassignments extends Ben_ltkpriorityassignmentsBase {
  ben_priorityname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  ben_priority?: object;
  _ben_priority_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}