// app/src/generated/models/Ben_ltkpriorityeventsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Ben_ltkpriorityeventsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Ben_ltkpriorityeventsstatecode = keyof typeof Ben_ltkpriorityeventsstatecode;
export const Ben_ltkpriorityeventsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Ben_ltkpriorityeventsstatuscode = keyof typeof Ben_ltkpriorityeventsstatuscode;

export interface Ben_ltkpriorityeventsBase {
  ben_actorid?: string;
  ben_actorname?: string;
  ben_at?: string;
  ben_detailjson?: string;
  ben_kind?: string;
  ben_ltkpriorityeventid: string;
  ben_name?: string;
  "ben_Priority@odata.bind"?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Ben_ltkpriorityeventsstatecode;
  statuscode?: Ben_ltkpriorityeventsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Ben_ltkpriorityevents extends Ben_ltkpriorityeventsBase {
  ben_priorityname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  ben_priority?: object;
  _ben_priority_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}